requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-yasg==1.21.7
phonenumbers==9.0.21
django-phonenumber-field==8.4.0
django-health-check==3.20.8
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.2.7
redis==5.0.1
Pillow==10.2.0
python-dateutil==2.8.2